refactor(#284): split MemoryPanel.tsx into sub-components
Extract file row, editor, preview, create form, search results, settings,
API hook, and utilities into memory-panel/ directory. MemoryPanel.tsx
reduced from 627 to 175 lines.
