+++ apps/desktop/src/components/settings/memory-panel/index.ts
@@ -1,0 +1,8 @@
+export { MemoryFileRow } from './MemoryFileRow';
+export { MemoryEditor } from './MemoryEditor';
+export { MemoryPreview } from './MemoryPreview';
+export { MemoryCreateForm } from './MemoryCreateForm';
+export { MemorySearchResults } from './MemorySearchResults';
+export { MemorySettings } from './MemorySettings';
+export { useMemoryApi } from './useMemoryApi';
+export { highlightMatch, formatBytesInline } from './memory-utils';
